So I did a minor optimization...
...which gives me just a couple perecents of performance advantage.
I might revert everything a day later, but before that I'm going to run
an "optimized" version and see the real result. To keep the reader
working at the same time, I have to make this commit.

And that's why I rarely do opensource. I hope this was an interesting
reading for you.